refactor : user service layer 리팩토링, test 수정(#520)
통합테스트 제거
