+++ backend/src/main/java/com/minibeit/user/service/AvatarService.java
@@ -1,16 +1,15 @@
 package com.minibeit.user.service;
 
-import com.minibeit.user.domain.Avatar;
 import com.minibeit.file.domain.S3Uploader;
-import com.minibeit.user.domain.repository.AvatarRepository;
+import com.minibeit.user.domain.Avatar;
+import com.minibeit.user.domain.User;
 import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
-@Component
+@Service
 @RequiredArgsConstructor
 public class AvatarService {
-    private final AvatarRepository avatarRepository;
     private final S3Uploader s3Uploader;
 
     public Avatar upload(MultipartFile file) {
@@ -20,9 +19,16 @@ public class AvatarService {
         return Avatar.create(s3Uploader.upload(file).toAvatar());
     }
 
+    public void update(Boolean avatarChanged, MultipartFile updatedAvatar, User findUser) {
+        if (avatarChanged) {
+            deleteOne(findUser.getAvatar());
+            Avatar avatar = upload(updatedAvatar);
+            findUser.updateAvatar(avatar);
+        }
+    }
+
     public void deleteOne(Avatar file) {
         if (file != null) {
-            avatarRepository.delete(file);
             s3Uploader.delete(file.getName());
         }
     }
