+++ backend/src/test/java/com/minibeit/auth/service/unit/RefreshTokenServiceUnitTest.java
@@ -4,7 +4,7 @@ import com.minibeit.auth.domain.TokenProvider;
 import com.minibeit.auth.service.RefreshTokenService;
 import com.minibeit.user.domain.repository.UserRepository;
 import com.minibeit.user.service.exception.UserNotFoundException;
-import com.minibeit.user.service.unit.MockUser;
+import com.minibeit.user.service.mock.MockUser;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
