+++ backend/src/test/java/com/minibeit/businessprofile/service/unit/BusinessProfileServiceUnitTest.java
@@ -7,12 +7,12 @@ import com.minibeit.businessprofile.service.BusinessProfileService;
 import com.minibeit.businessprofile.service.dto.BusinessProfileResponse;
 import com.minibeit.businessprofile.service.exception.BusinessProfileNotFoundException;
 import com.minibeit.businessprofile.service.exception.UserBusinessProfileNotFoundException;
-import com.minibeit.file.avatar.service.unit.MockFile;
+import com.minibeit.file.mock.MockFile;
 import com.minibeit.user.domain.Avatar;
 import com.minibeit.user.service.AvatarService;
 import com.minibeit.user.domain.repository.UserRepository;
 import com.minibeit.user.service.exception.UserNotFoundException;
-import com.minibeit.user.service.unit.MockUser;
+import com.minibeit.user.service.mock.MockUser;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -23,7 +23,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import java.util.Optional;
 
 import static com.minibeit.businessprofile.service.unit.MockBusinessProfile.BusinessProfile1.*;
-import static com.minibeit.user.service.unit.MockUser.MockUser1;
+import static com.minibeit.user.service.mock.MockUser.MockUser1;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
