+++ backend/src/test/java/com/minibeit/message/service/unit/SmsServiceUnitTest.java
@@ -6,7 +6,7 @@ import com.minibeit.message.service.component.SmsProps;
 import com.minibeit.user.domain.repository.UserRepository;
 import com.minibeit.user.domain.repository.UserVerificationCodeRepository;
 import com.minibeit.user.service.exception.UserNotFoundException;
-import com.minibeit.user.service.unit.MockUser;
+import com.minibeit.user.service.mock.MockUser;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
