+++ backend/src/test/java/com/minibeit/post/service/unit/PostApplicantServiceUnitTest.java
@@ -9,8 +9,7 @@ import com.minibeit.post.domain.repository.PostLikeRepository;
 import com.minibeit.post.service.PostApplicantService;
 import com.minibeit.post.service.exception.PostApplicantNotFoundException;
 import com.minibeit.post.service.exception.PostDoDateNotFoundException;
-import com.minibeit.post.service.unit.mock.MockPostApplicant;
-import com.minibeit.user.service.unit.MockUser;
+import com.minibeit.user.service.mock.MockUser;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
