+++ backend/src/test/java/com/minibeit/post/service/unit/RejectPostServiceUnitTest.java
@@ -3,7 +3,7 @@ package com.minibeit.post.service.unit;
 import com.minibeit.post.domain.repository.RejectPostRepository;
 import com.minibeit.post.service.RejectPostService;
 import com.minibeit.post.service.unit.mock.MockPageDto;
-import com.minibeit.user.service.unit.MockUser;
+import com.minibeit.user.service.mock.MockUser;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
