+++ backend/src/test/java/com/minibeit/post/service/unit/mock/MockRejectPost.java
@@ -4,7 +4,7 @@ import com.minibeit.businessprofile.domain.BusinessProfile;
 import com.minibeit.businessprofile.service.unit.MockBusinessProfile;
 import com.minibeit.post.domain.RejectPost;
 import com.minibeit.user.domain.User;
-import com.minibeit.user.service.unit.MockUser;
+import com.minibeit.user.service.mock.MockUser;
 
 import java.time.LocalDateTime;
