+++ backend/src/test/java/com/minibeit/review/service/unit/BusinessUserReviewServiceUnitTest.java
@@ -14,7 +14,7 @@ import com.minibeit.review.service.BusinessUserReviewService;
 import com.minibeit.review.service.exception.BusinessReviewDetailNotFoundException;
 import com.minibeit.user.domain.repository.UserRepository;
 import com.minibeit.user.service.exception.UserNotFoundException;
-import com.minibeit.user.service.unit.MockUser;
+import com.minibeit.user.service.mock.MockUser;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
