+++ backend/src/test/java/com/minibeit/user/service/UserServiceUnitTest.java
@@ -1,14 +1,11 @@
-package com.minibeit.user.service.unit;
+package com.minibeit.user.service;
 
 import com.minibeit.businessprofile.service.unit.MockBusinessProfile;
-import com.minibeit.file.avatar.service.unit.MockFile;
-import com.minibeit.user.service.AvatarService;
 import com.minibeit.school.service.integrate.Schools;
 import com.minibeit.school.service.unit.MockSchool;
 import com.minibeit.user.domain.UserValidator;
 import com.minibeit.user.domain.repository.UserRepository;
 import com.minibeit.user.domain.repository.UserVerificationCodeRepository;
-import com.minibeit.user.service.UserService;
 import com.minibeit.user.service.dto.UserResponse;
 import com.minibeit.user.service.exception.UserNotFoundException;
 import com.minibeit.user.service.exception.UserVerificationCodeNotFoundException;
@@ -22,7 +19,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import java.util.Optional;
 
 import static com.minibeit.businessprofile.service.unit.MockBusinessProfile.BusinessProfile1.ID;
-import static com.minibeit.user.service.unit.MockUser.MockUser1.*;
+import static com.minibeit.user.service.mock.MockUser.MockUser1.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
@@ -36,11 +33,11 @@ public class UserServiceUnitTest {
     @Mock
     UserRepository userRepository;
     @Mock
+    UserVerificationCodeRepository userVerificationCodeRepository;
+    @Mock
     Schools schools;
     @Mock
     AvatarService avatarService;
-    @Mock
-    UserVerificationCodeRepository userVerificationCodeRepository;
     @Mock
     UserValidator userValidator;
     @InjectMocks
@@ -58,6 +55,7 @@ public class UserServiceUnitTest {
         assertThat(response.getNickname()).isEqualTo(SIGNUP_REQUEST.getNickname());
         assertThat(response.getSchoolId()).isEqualTo(SIGNUP_REQUEST.getSchoolId());
         verify(userValidator).nicknameValidate(any());
+        verify(avatarService).upload(any());
     }
 
     @Test
@@ -72,15 +70,12 @@ public class UserServiceUnitTest {
     @DisplayName("개인정보 수정 성공")
     public void update() {
         given(userRepository.findByIdWithAvatar(any())).willReturn(Optional.of(USER));
-        given(schools.getOne(SCHOOL_ID)).willReturn(MockSchool.School1.SCHOOL);
-        given(avatarService.upload(any())).willReturn(MockFile.MockFile1.AVATAR);
+        given(schools.getOne(any())).willReturn(MockSchool.School1.SCHOOL);
 
-        UserResponse.CreateOrUpdate response = userService.update(UPDATE_REQUEST, USER);
+        userService.update(UPDATE_REQUEST, USER);
 
-        assertThat(response.getNickname()).isEqualTo(UPDATED_NICKNAME);
-        verify(avatarService).deleteOne(any());
-        verify(avatarService).upload(any());
-        verify(userValidator).nicknameValidate(any());
+        verify(userValidator).updateValidate(any(), any());
+        verify(avatarService).update(any(), any(),any());
     }
 
     @Test
@@ -155,6 +150,8 @@ public class UserServiceUnitTest {
     @Test
     @DisplayName("회원 탈퇴 성공")
     public void deleteOne() {
+        given(userRepository.findByIdWithAvatar(any())).willReturn(Optional.of(USER));
+
         userService.deleteOne(USER);
 
         verify(userValidator).deleteValidate(any());
